attempt at VRS integration into genomicVariations  default schema
This is an attempt to allow the use of VRS variation classes, while also keeping an updated version of the current Beacon schema (using VRS location, which was agreed upon) as an option.
All the additional properties beyond the definition of the variant itself (annotations, caseLevel data etc.) are kept in place.

The change here introduces the root level `variation` property but other options could be possible.