filterIds in GET requests
So far the Beacon models do not contain an option to insert filters into GET requests. This PR adds the parameter `filterIds` to the endpoints. The naming reflects the use for the `id` values of the `filteringTerms` schema.

This addition addresses the majority of `filteringTerms` use cases (while not addressing filter specific modifications such as per-filter descendant term yes/no).

Note: The Progenetix resource has used this format since ~2018 for all its interface & API interactions.