Adding minimal requestParameters schemas to all entry types
As discussed in https://github.com/ga4gh-beacon/beacon-v2-Models/issues/50 it seems cleaner to have requestParameters defined for all entry types which at the minimum should contain the `id` parameter.

This PR add the documents as well as parameters for each `id` value defined in the schema (e.g. for `biosample` this would be `id` and `individualId`).